tighter infobars camera reset function bullets hit walls / no longer shoot through them
git-svn-id: svn://84.194.112.164/bengine@220 2ee6282c-0497-4b59-8b7c-bcac2105c061

+++ src/utils/infobar.h
@@ -12,7 +12,7 @@ class Infobar
 		Infobar();
 		~Infobar();
 
-		float barheight;
+		unsigned int barheight;
 
 		void draw();
 		void swap();
@@ -22,8 +22,8 @@ class Infobar
 		Fps fps;
 
 		// horizontal/vertical spacers (padding)
-		float hsp;
-		float vsp;
+		unsigned int hsp;
+		unsigned int vsp;
 
 		bool active;
 };
